LayoutTests/fast/regex/parentheses-expected.txt

This page tests handling of parentheses subexpressions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS regexp1.exec('abc') is ['ab','a','b']
PASS regexp2.exec('abacadabe') is ['abe','ab','b','b']
PASS regexp3.exec('abacadabe') is ['abe','ab','b',undefined]
PASS regexp4.exec('abacadabe') is ['abe','ab','b',undefined]
PASS regexp5.exec('abacadabe') is ['abe','ab','b','b',undefined,undefined]
PASS regexp6.exec('abcde') is ['ab','ab','b','b',undefined,undefined]
PASS regexp7.exec('abc') is ['abc','ab','b']
PASS regexp8.exec('bcaddxqy') is ['qy','q','q','y']
PASS regexp9.exec('asdfjejgsdflaksdfjkeljghkjea') is ['a','a',undefined]
PASS regexp10.exec('asdfjejgsdflaksdfjkeljghat') is ['at']
PASS regexp11.exec('Developing with JavaScript is dangerous, do not try it without assistance') is null
PASS regexp12.exec('Seattle, WA to Buckley, WA') is ['Seattle, WA to Buckley, WA', undefined, 'Seattle', 'WA', undefined, 'Buckley', 'WA']
PASS regexp13.exec('zxcasd;fl ^AaaAAaaaf;lrlrzs') is ['AaaAAaaaf;lrlrzs',undefined,'AaaAAaaaf;lrlrzs']
PASS regexp14.exec('b') is ['b',undefined,'b']
PASS regexp15.exec('abdf') is ['abdf',undefined,'abd','f']
PASS regexp16.exec('abc') is ['ab','a','b']
PASS regexp17.exec('bcaDxqy') is ['Dx','D']
PASS regexp18.exec('Hello: World') is ['Hello:',':']
PASS regexp19.exec('barrel') is ['bar','']
PASS regexp20.exec('barrel') is ['barrel','rel']
PASS regexp20.exec('2barrel') is ['2barrel','rel']
PASS regexp21.exec('abc') is ['ab','ab','b']
PASS regexp22.exec('abcdlskfgjdslkfg') is null
PASS regexp23.exec('<html xmlns="http://www.w3.org/1999/xhtml"') is ['"http://www.w3.org/1999/xhtml"']
PASS regexp24.exec('123') is null
PASS regexp25.exec('this is a test') is ['this','this',undefined]
PASS regexp25.exec('!this is a test') is null
PASS 'Hi Bob'.match(/(Rob)|(Bob)|(Robert)|(Bobby)/) is ['Bob',undefined,'Bob',undefined,undefined]
PASS successfullyParsed is true

TEST COMPLETE
